# target/classes/swaggerapi.yml
swagger: '2.0'
info:
  title: SpringBoot Test API
  description: |
    Spring boot API services for learning
  version: 0.01.5
host: localhost
schemes:
  - https
basePath: /api/services
consumes:
  - application/json
produces:
  - application/json
...
paths:
...
  /blogs/UpdateBlog:
    put:
      parameters:
        - in: body
          name: Blog
          description: Blog details
          required: true
          schema:
            $ref: '#/definitions/Blog'
      summary: update Blog details
      deprecated: false
      responses:
        200:
          description: Success
        401:
          description: Not authenticated. All activities are logged per user, so users must be authenticated before they can call this endpoint.
        403:
          description: Not authorized for any data. Note this is not the case if user has a role of Administrator or Manager.
        404:
          description: The origin server did not find a current representation for the target resource or is not willing to disclose that one exists.
        405:
          description: The method received in the request-line is known by the origin server but not supported by the target resource.
        500:
          description: The server encountered an unexpected condition that prevented it from fulfilling the request.
        503:
          description: The server is currently unable to handle the request due to a temporary overload or scheduled maintenance, which will likely be alleviated after some delay
        default:
          description: Unexpected ErrorResponse
...
  /blogs/addBlog:
    post:
      parameters:
        - in: body
          name: Blog
          description: Blog details
          required: true
          schema:
            $ref: '#/definitions/Blog'
      summary: add user details
      deprecated: false
      responses:
        200:
          description: Success
        401:
          description: Not authenticated. All activities are logged per user, so users must be authenticated before they can call this endpoint.
        403:
          description: Not authorized for any data. Note this is not the case if user has a role of Administrator or Manager.
        404:
          description: The origin server did not find a current representation for the target resource or is not willing to disclose that one exists.
        405:
          description: The method received in the request-line is known by the origin server but not supported by the target resource.
        500:
          description: The server encountered an unexpected condition that prevented it from fulfilling the request.
        503:
          description: The server is currently unable to handle the request due to a temporary overload or scheduled maintenance, which will likely be alleviated after some delay
        default:
          description: Unexpected ErrorResponse
  /blogs:
    get:
      summary: Get blog details
      deprecated: false
      responses:
        200:
          description: Success
          schema:
            $ref: '#/definitions/Blogs'
        401:
          description: Not authenticated. All activities are logged per user, so users must be authenticated before they can call this endpoint.
        403:
          description: Not authorized for any data. Note this is not the case if user has a role of Administrator or Manager.
        404:
          description: The origin server did not find a current representation for the target resource or is not willing to disclose that one exists.
        405:
          description: The method received in the request-line is known by the origin server but not supported by the target resource.
        500:
          description: The server encountered an unexpected condition that prevented it from fulfilling the request.
        503:
          description: The server is currently unable to handle the request due to a temporary overload or scheduled maintenance, which will likely be alleviated after some delay
        default:
          description: Unexpected ErrorResponse
  /blogs/deleteById:
    delete:
      parameters:
        - in: query
          name: blogId
          type: integer
          description: Delete Blog By Id
          required: true
          allowEmptyValue: false
      summary: delete blog By Id
      deprecated: false
      responses:
        200:
          description: Success
        401:
          description: Not authenticated. All activities are logged per user, so users must be authenticated before they can call this endpoint.
        403:
          description: Not authorized for any data. Note this is not the case if user has a role of Administrator or Manager.
        404:
          description: The origin server did not find a current representation for the target resource or is not willing to disclose that one exists.
        405:
          description: The method received in the request-line is known by the origin server but not supported by the target resource.
        500:
          description: The server encountered an unexpected condition that prevented it from fulfilling the request.
        503:
          description: The server is currently unable to handle the request due to a temporary overload or scheduled maintenance, which will likely be alleviated after some delay
        default:
          description: Unexpected ErrorResponse
  /blogs/getBlogById:
    get:
      parameters:
        - in: query
          description: blog id
          name: blogId
          type: integer
          required: true
          allowEmptyValue: false
      summary: Get Blog details By Id
      deprecated: false
      responses:
        200:
          description: Success
          schema:
            $ref: '#/definitions/Blog'
        401:
          description: Not authenticated. All activities are logged per user, so users must be authenticated before they can call this endpoint.
        403:
          description: Not authorized for any data. Note this is not the case if user has a role of Administrator or Manager.
        404:
          description: The origin server did not find a current representation for the target resource or is not willing to disclose that one exists.
        405:
          description: The method received in the request-line is known by the origin server but not supported by the target resource.
        500:
          description: The server encountered an unexpected condition that prevented it from fulfilling the request.
        503:
          description: The server is currently unable to handle the request due to a temporary overload or scheduled maintenance, which will likely be alleviated after some delay
        default:
          description: Unexpected ErrorResponse
  /blogs/getAllBlogByUserId:
    get:
      parameters:
        - in: query
          description: userBlog Id
          name: userId
          type: integer
          required: true
          allowEmptyValue: false
      summary: Get All blog by userid
      deprecated: false
      responses:
        200:
          description: Success
          schema:
            $ref: '#/definitions/Blogs'
        401:
          description: Not authenticated. All activities are logged per user, so users must be authenticated before they can call this endpoint.
        403:
          description: Not authorized for any data. Note this is not the case if user has a role of Administrator or Manager.
        404:
          description: The origin server did not find a current representation for the target resource or is not willing to disclose that one exists.
        405:
          description: The method received in the request-line is known by the origin server but not supported by the target resource.
        500:
          description: The server encountered an unexpected condition that prevented it from fulfilling the request.
        503:
          description: The server is currently unable to handle the request due to a temporary overload or scheduled maintenance, which will likely be alleviated after some delay
        default:
          description: Unexpected ErrorResponse
...
definitions:
...
  Blogs:
    type: array
    title: All Blog Details
    items:
      $ref: '#/definitions/Blog'
  Blog:
    type: object
    properties:
      blogId:
        type: integer
        description: Blog blog_id
      blogtitle:
        type: string
        description: Blog blog_title
      blogbody:
        type: string
        description: Blog blog_body
      publishby:
        type: string
        description: Blog publish_by
      isdeleted:
        type: string
        description: Blog is_deleted
      createdon:
        type: string
        format: timestamp
        description: deletion date and time
        example: "2021-01-30T08:30:00Z"
      updatedon:
        type: string
        format: timestamp
        description: deletion date and time
        example: "2021-01-30T08:30:00Z"
      userBlogId:
        type: integer
        description: userBlog Id
...